+++ cmd/hidden_lake/service/internal/handler/handler_test.go
@@ -12,12 +12,13 @@ import (
 	"github.com/number571/go-peer/internal/interrupt"
 	"github.com/number571/go-peer/pkg/client"
 	"github.com/number571/go-peer/pkg/client/message"
-	"github.com/number571/go-peer/pkg/client/queue"
 	"github.com/number571/go-peer/pkg/crypto/asymmetric"
 	"github.com/number571/go-peer/pkg/logger"
 	"github.com/number571/go-peer/pkg/network"
 	"github.com/number571/go-peer/pkg/network/anonymity"
+	"github.com/number571/go-peer/pkg/network/anonymity/queue"
 	"github.com/number571/go-peer/pkg/network/conn"
+	net_message "github.com/number571/go-peer/pkg/network/message"
 	"github.com/number571/go-peer/pkg/queue_set"
 	"github.com/number571/go-peer/pkg/storage"
 	"github.com/number571/go-peer/pkg/storage/database"
@@ -181,11 +182,12 @@ func testNewNode(dbPath, addr string) anonymity.INode {
 	if err != nil {
 		return nil
 	}
+	networkMask := uint64(1)
 	node := anonymity.NewNode(
 		anonymity.NewSettings(&anonymity.SSettings{
 			FServiceName:   "TEST",
 			FRetryEnqueue:  0,
-			FNetworkMask:   1,
+			FNetworkMask:   networkMask,
 			FFetchTimeWait: time.Minute,
 		}),
 		logger.NewLogger(
@@ -207,6 +209,11 @@ func testNewNode(dbPath, addr string) anonymity.INode {
 				}),
 				asymmetric.LoadRSAPrivKey(testutils.Tc1PrivKey1024),
 			),
+			func() (uint64, net_message.ISettings) {
+				return networkMask, net_message.NewSettings(&net_message.SSettings{
+					FWorkSizeBits: testutils.TCWorkSize,
+				})
+			},
 		),
 		asymmetric.NewListPubKeys(),
 	)