+++ pkg/network/anonymity/examples/ping-ping/main.go
@@ -10,14 +10,15 @@ import (
 
 	"github.com/number571/go-peer/pkg/client"
 	"github.com/number571/go-peer/pkg/client/message"
-	"github.com/number571/go-peer/pkg/client/queue"
 	"github.com/number571/go-peer/pkg/crypto/asymmetric"
 	"github.com/number571/go-peer/pkg/logger"
 	"github.com/number571/go-peer/pkg/network"
 	"github.com/number571/go-peer/pkg/network/anonymity"
 	"github.com/number571/go-peer/pkg/network/anonymity/adapters"
 	anon_logger "github.com/number571/go-peer/pkg/network/anonymity/logger"
+	"github.com/number571/go-peer/pkg/network/anonymity/queue"
 	"github.com/number571/go-peer/pkg/network/conn"
+	net_message "github.com/number571/go-peer/pkg/network/message"
 	"github.com/number571/go-peer/pkg/queue_set"
 	"github.com/number571/go-peer/pkg/storage"
 	"github.com/number571/go-peer/pkg/storage/database"
@@ -126,11 +127,12 @@ func newNode(serviceAddress, name, dbPath string) anonymity.INode {
 	if err != nil {
 		return nil
 	}
+	networkMask := uint64(1)
 	return anonymity.NewNode(
 		anonymity.NewSettings(&anonymity.SSettings{
 			FServiceName:   name,
 			FRetryEnqueue:  0,
-			FNetworkMask:   1,
+			FNetworkMask:   networkMask,
 			FFetchTimeWait: time.Minute,
 		}),
 		logger.NewLogger(
@@ -177,6 +179,11 @@ func newNode(serviceAddress, name, dbPath string) anonymity.INode {
 				}),
 				asymmetric.NewRSAPrivKey(keySize),
 			),
+			func() (uint64, net_message.ISettings) {
+				return networkMask, net_message.NewSettings(&net_message.SSettings{
+					FWorkSizeBits: workSize,
+				})
+			},
 		),
 		asymmetric.NewListPubKeys(),
 	)