+++ pkg/network/anonymity/queue/doc.go
@@ -1,4 +1,4 @@
-// Package queue allows you to queue and retrieve encrypted messages from the queue.
+// Package queue allows you to queue and retrieve network encrypted messages from the queue.
 //
 // The internal process of the queue generates constantly false messages for the
 // continuous possibility of receiving messages.