+++ pkg/network/anonymity/queue/types.go
@@ -5,17 +5,21 @@ import (
 
 	"github.com/number571/go-peer/pkg/client"
 	"github.com/number571/go-peer/pkg/client/message"
+	net_message "github.com/number571/go-peer/pkg/network/message"
 	"github.com/number571/go-peer/pkg/types"
 )
 
+type INetworkSettingsFunc func() (uint64, net_message.ISettings)
+
 type IMessageQueue interface {
 	types.IApp
+	ClearQueue()
 
 	GetSettings() ISettings
 	GetClient() client.IClient
 
 	EnqueueMessage(message.IMessage) error
-	DequeueMessage() <-chan message.IMessage
+	DequeueMessage() <-chan net_message.IMessage
 }
 
 type ISettings interface {